+++ src/utils/to-inline-style.js
@@ -1,10 +1,10 @@
 const camelToDash = require('./camel-to-dash');
 
-function toInlineStyle(obj) {
+function toInlineStyle(obj, withStyle = true) {
     obj = Object.entries(obj).reduce((styleString, [propName, propValue]) => {
         return `${styleString}${camelToDash(propName)}:${propValue};`;
     }, '');
-    return `style="${obj}"`
+    return withStyle ? `style="${obj}"` : obj;
 }
 
 module.exports = toInlineStyle;