+++ CHANGELOG.md
@@ -5,9 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
-- **Added** new property `appRoot` to component instance
 - **Breaking changes**
     - Refactored `style` to `toStyle` method
+- **Added** new property `appRoot` to component instance
+- **Added** new property `style` to component instance
 
 ## [1.3.5] - 2018-07-11
 - **Fixed** attributes was removed with falsy value, now only if empty